Text Fix for Login Header
Not graceful, but the DT Getai font adds something like an internal padding at the bottom. No way to fix that I know of other than adding 20px margin to top.

Fixed header text, size, and tagline font.

Next need to make it so the login pop-ups are part of flow and push header up.