Starting Search Functionality
When focusing the searchbar (or when there is no input): display the "Create Patient/Task" buttons which will open their respective modals. This works.

Then, when typing input, get %ILIKE% results for relevant patients, tasks, or even pages. This will hopefully make it easier to navigate to specific content. We'll probably need content anchors too so a search for "Jared Carthalien" will display the Patients page with Jared front and center.
This needs work before it is complete.